requirements.txt
Django
djangorestframework
pillow
pandas
numpy
pytest-django
# python -m pytest

